Fix multiple versions of pkgs being installed
Multiple versions of numpy and scikit-image were installed because
`python setup.py develop` does not uninstall old versions before
installing a newer version. Using `pip install -e .` fixes this issue.

The `--no-use-pep517` option fixes the issue of not being able to import
neuralpredictors after installation. Without this option pip would install
it into /usr/lib/python3.8/site-packages/ which was not in PYTHONPATH.
With this option pip installs it into
/usr/local/lib/python3.8/dist-packages which is in PYTHONPATH. This
issue did not arise for the data_port package because unlike
neuralpredictors it does not have a pyproject.toml file.